@@ -15,7 +15,7 @@
         /// <param name="history">The history of object data to merge</param>
         /// <param name="historyComparer">The comparer instance to use for sorting the history before merge. Will
         /// default to the type's default comparer if none specified.</param>
-        /// <param name="nonDefaultMergeCriteria">Any object/property-specific criteria where by to determine
+        /// <param name="nonDefaultMergeCriteria">Any object/property-specific criteria whereby to determine
         /// when to overwrite the value when evaluating the property's history.</param>
         /// <returns>A single object instance containing the most recent data.</returns>
         /// <typeparam name="T">The type of object to merge.</typeparam>
@@ -29,7 +29,7 @@
         /// </summary>
         /// <param name="historyComparer">The comparer instance to use for sorting the history before merge. Will
         /// default to the type's default comparer if none specified.</param>
-        /// <param name="nonDefaultMergeCriteria">Any object/property-specific criteria where by to determine
+        /// <param name="nonDefaultMergeCriteria">Any object/property-specific criteria whereby to determine
         /// when to overwrite the value when evaluating the property's history.</param>
         /// <param name="history">The history of object data to merge, in order.</param>
         /// <returns>A single object instance containing the most recent data.</returns>
