@@ -12,7 +12,7 @@
             //                    .Append(nameof(Model.Prop1))
             //                    .Append(nameof(Model.Prop2));
             //string actual = BreadcrumbHelper<Model>.Of(a => a.Prop1.Prop2);
-            //Assert.AreEqual(expected, actual);
+            //Assert.That(expected, actual);
         }
     }
 }
